Move to Github Actions
We had an issue where the default Github build started behaving
differently, so we'd like to pin our versions. This commit will
have Github Actions do the build against our Gemfile.lock.

Note that the effects of this commit on the actual deployed website
*will not take effect* until the Github settings are changed to
reflect that the site is now hosted on the gh-pages branch.